requirements.txt
flask
requests
beautifulsoup4
openai
python-dotenv
Werkzeug>=2.0
metalpriceapi
gunicorn
